# apps/service-gateway/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.SysBusinessPartner:
    properties:
      bpCode:
        type: string
      bpName:
        type: string
      bpType:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      extData:
        items:
          type: integer
        type: array
      id:
        type: integer
      taxID:
        type: string
      updatedAt:
        type: string
    type: object
  models.SysExpenseClaim:
    properties:
      amount:
        type: number
      claimNo:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      employeeID:
        type: string
      extData:
        items:
          type: integer
        type: array
      id:
        type: integer
      reason:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.SysGLAccount:
    properties:
      accountCode:
        type: string
      accountName:
        type: string
      accountType:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      extData:
        items:
          type: integer
        type: array
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.SysOrganization:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      extData:
        items:
          type: integer
        type: array
      id:
        type: integer
      orgCode:
        type: string
      orgName:
        type: string
      orgType:
        type: string
      parentID:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个高度可扩展的开源 ERP 底座，支持动态字段(JSONB)、复杂审批流以及物料需求计划(MRP)。
  title: OpenERP 企业级信息化中枢 API
  version: "1.0"
paths:
  /openerp/v1/business-partners:
    get:
      consumes:
      - application/json
      description: 获取业务伙伴列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取业务伙伴列表
      tags:
      - 业务伙伴管理
    post:
      consumes:
      - application/json
      description: 创建一个新的业务伙伴（客户或供应商）
      parameters:
      - description: 业务伙伴数据
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/models.SysBusinessPartner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 创建业务伙伴
      tags:
      - 业务伙伴管理
  /openerp/v1/business-partners/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除业务伙伴
      parameters:
      - description: 业务伙伴ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除业务伙伴
      tags:
      - 业务伙伴管理
    put:
      consumes:
      - application/json
      description: 更新业务伙伴数据
      parameters:
      - description: 业务伙伴ID
        in: path
        name: id
        required: true
        type: integer
      - description: 业务伙伴数据
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/models.SysBusinessPartner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新业务伙伴
      tags:
      - 业务伙伴管理
  /openerp/v1/expense-claims:
    get:
      description: 获取所有的费用报销单
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取费用报销单列表
      tags:
      - ExpenseClaim
    post:
      consumes:
      - application/json
      description: 创建一个新的费用报销单
      parameters:
      - description: 报销单信息
        in: body
        name: claim
        required: true
        schema:
          $ref: '#/definitions/models.SysExpenseClaim'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 创建费用报销单
      tags:
      - ExpenseClaim
  /openerp/v1/expense-claims/{id}:
    delete:
      description: 删除指定的费用报销单
      parameters:
      - description: 报销单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除费用报销单
      tags:
      - ExpenseClaim
    put:
      consumes:
      - application/json
      description: 更新现有的费用报销单
      parameters:
      - description: 报销单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 报销单信息
        in: body
        name: claim
        required: true
        schema:
          $ref: '#/definitions/models.SysExpenseClaim'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新费用报销单
      tags:
      - ExpenseClaim
...
  /openerp/v1/gl-accounts:
    get:
      description: 获取所有的总账科目
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取总账科目列表
      tags:
      - GLAccount
    post:
      consumes:
      - application/json
      description: 创建一个新的总账科目
      parameters:
      - description: 总账科目信息
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.SysGLAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 创建总账科目
      tags:
      - GLAccount
  /openerp/v1/gl-accounts/{id}:
    delete:
      description: 删除指定的总账科目
      parameters:
      - description: 科目ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除总账科目
      tags:
      - GLAccount
    put:
      consumes:
      - application/json
      description: 更新现有的总账科目
      parameters:
      - description: 科目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 总账科目信息
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.SysGLAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新总账科目
      tags:
      - GLAccount
...
  /openerp/v1/organizations:
    get:
      description: 获取组织架构列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取组织架构列表
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: 创建一个新的组织架构节点
      parameters:
      - description: 组织架构信息
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/models.SysOrganization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 创建组织架构
      tags:
      - Organization
  /openerp/v1/organizations/{id}:
    delete:
      description: 删除指定的组织架构节点
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除组织架构
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: 更新现有的组织架构节点
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      - description: 组织架构信息
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/models.SysOrganization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新组织架构
      tags:
      - Organization
...
swagger: "2.0"